LayoutTests/crypto/subtle/aes-kw-generate-export-raw-key-expected.txt

Test exporting an AES-KW key with raw format

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Generating a key...
Exporting a key...
PASS key.byteLength is 32
PASS successfullyParsed is true

TEST COMPLETE
